incremental progress on frontend
