+++ app/src/App.css
@@ -1,3 +1,4 @@
 :root {
   --selection-color: #9792e3;
+  --nav-bar-height: 68px;
 }
