+++ app/src/Chart.css
@@ -2,8 +2,9 @@
   display: flex;
   justify-content: center;
   align-items: center;
+  text-align: center;
   flex-direction: row;
-  height: 92vh;
+  height: 100%;
   width: 100%;
   gap: 2px;
 }
@@ -30,6 +31,15 @@
   line-height: 1;
   grid-row: 4;
 }
+.chart-tab {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100vw;
+  top: 72px;
+  height: 94vh;
+}
 
 .timeline-button {
   color: red;
@@ -39,7 +49,6 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  width: "auto";
 }
 .options-bar {
   display: flex;
@@ -50,7 +59,12 @@
 .flex-row {
   display: flex;
   flex-direction: row;
+  justify-content: center;
+  align-items: center;
 }
 .chart-download-button {
   align-items: flex-end;
+}
+.chart-container {
+  height: calc(100vh - var(--nav-bar-height));
 }
