+++ app/src/crossplot/CrossPlotChart.module.css
@@ -1,0 +1,4 @@
+.container {
+  display: grid;
+  grid-template-columns: 26.5vw 1fr;
+}
