+++ app/src/crossplot/CrossPlotSideBar.module.css
@@ -1,0 +1,9 @@
+.crossPlotSideBar {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: calc(100vh - var(--nav-bar-height));
+  background-color: #f5f5f5;
+  border-right: 1px solid #e0e0e0;
+  overflow-y: auto;
+}
