+++ app/src/util/constant.ts
@@ -14,3 +14,5 @@ export const defaultColors: string[] = [
 export const snackbarTextLengthLimit: number = 70;
 
 export const altUnitNamePrefix: string = "Chart Title in ";
+
+export const navBarHeight = "72px";
